Refactoring validateDockerfile to dockerfile.go and change func name
